Move converter functions to separate modules This is to avoid breaking the max-line-length rule, which is set at 500 lines.
